refactor(db): extract db/core.js + add export-manifest gate — decomposition Wave 0
First wave of the db.js decomposition (4499-line monolith -> server/db/ modules +
a thin barrel). Behavior-preserving: no consumer's `import from '../db.js'` changes.

- server/db/core.js: the foundation module — the live `db` binding (assigned only
  in initDBConnection; ESM live-binding, so siblings read bare `db` verbatim), plus
  initDBConnection (connection + PRAGMAs + migrations + schema + instance_config
  seed), the stmt() prepared-statement cache, buildUpdate() generic UPDATE builder,
  and getDB(). Bodies moved verbatim from db.js.
- server/db.js: now imports { db, stmt, buildUpdate, initDBConnection, getDB, DB_PATH }
  from ./db/core.js; all other functions stay (later waves) and reach the shared
  bindings via the ESM live binding. -245/+3.

Gate (the invariant that makes the whole campaign safe):
- test/refactor/db-manifest.{mjs,test.js,snapshot}: pins db.js's public surface at
  exactly 308 exports (name:typeof:arity). Any lost/added/re-typed/re-arity'd export
  fails `node db-manifest.mjs --check` AND the vitest suite.

Verification: manifest --check green (308 unchanged); full suite 310/310.

Executor: GLM-5.2 (agentic, gate-driven). Reviewed + count-baselined + verified by m5Max.
